Update libgit2 to 3f8d005
https://github.com/libgit2/libgit2/compare/e0383fa...3f8d005

+++ LibGit2Sharp/Core/GitDiff.cs
@@ -205,8 +205,8 @@ namespace LibGit2Sharp.Core
 
         /* options controlling how to diff text is generated */
 
-        public ushort ContextLines;
-        public ushort InterhunkLines;
+        public uint ContextLines;
+        public uint InterhunkLines;
         public ushort IdAbbrev;
         public Int64 MaxSize;
         public IntPtr OldPrefixString;
